LayoutTests/platform/win/fast/events/panScroll-correct-direction-expected.txt

CONSOLE MESSAGE: line 14: Mousedown with 1
This tests that pan scrolling scrolls the proper direction without event.preventDefault and the middle button mousedown event is fired. To test manually, try to pan scroll sideways inside this div.

Success! Div with overflow was scrolled correctly
